+++ jason/src/rpc/heartbeat.rs
@@ -1,4 +1,4 @@
-use std::{cell::RefCell, convert::From, rc::Rc};
+use std::{cell::RefCell, rc::Rc};
 
 use derive_more::{Display, From};
 use medea_client_api_proto::ClientMsg;
@@ -6,7 +6,7 @@ use tracerr::Traced;
 use wasm_bindgen::{prelude::*, JsCast};
 
 use crate::{
-    rpc::websocket::{SocketError, WebSocket},
+    rpc::{RpcTransport, TransportError},
     utils::{window, IntervalHandle, JsCaused, JsError},
 };
 
@@ -14,8 +14,8 @@ use crate::{
 #[derive(Debug, Display, From, JsCaused)]
 #[allow(clippy::module_name_repetitions)]
 pub enum HeartbeatError {
-    /// Occurs when `ping` cannot be send because no socket.
-    #[display(fmt = "unable to ping: no socket")]
+    /// Occurs when `ping` cannot be send because no transport.
+    #[display(fmt = "unable to ping: no transport")]
     NoSocket,
 
     /// Occurs when a handler cannot be set to send `ping`.
@@ -24,7 +24,7 @@ pub enum HeartbeatError {
 
     /// Occurs when socket failed to send `ping`.
     #[display(fmt = "failed to send ping: {}", _0)]
-    SendPing(#[js(cause)] SocketError),
+    SendPing(#[js(cause)] TransportError),
 }
 
 type Result<T> = std::result::Result<T, Traced<HeartbeatError>>;
@@ -40,21 +40,21 @@ struct InnerHeartbeat {
     num: u64,
     /// Timestamp of last pong received.
     pong_at: Option<f64>,
-    /// WebSocket connection with remote server.
-    socket: Option<Rc<WebSocket>>,
+    /// Connection with remote RPC server.
+    transport: Option<Rc<dyn RpcTransport>>,
     /// Handler of sending `ping` task. Task is dropped if you drop handler.
     ping_task: Option<PingTaskHandler>,
 }
 
 impl InnerHeartbeat {
-    /// Send ping message into socket.
-    /// Returns error no open socket.
+    /// Send ping message to RPC server.
+    /// Returns errors if no open transport found.
     fn send_now(&mut self) -> Result<()> {
-        match self.socket.as_ref() {
+        match self.transport.as_ref() {
             None => Err(tracerr::new!(HeartbeatError::NoSocket)),
-            Some(socket) => {
+            Some(transport) => {
                 self.num += 1;
-                Ok(socket
+                Ok(transport
                     .send(&ClientMsg::Ping(self.num))
                     .map_err(tracerr::map_from_and_wrap!())?)
             }
@@ -70,26 +70,26 @@ struct PingTaskHandler {
 
 impl Heartbeat {
     /// Returns new instance of [`interval`] with given interval for ping in
-    /// seconds.
+    /// milliseconds.
     pub fn new(interval: i32) -> Self {
         Self(Rc::new(RefCell::new(InnerHeartbeat {
             interval,
             num: 0,
             pong_at: None,
-            socket: None,
+            transport: None,
             ping_task: None,
         })))
     }
 
-    /// Starts [`Heartbeat`] for given [`WebSocket`].
+    /// Starts [`Heartbeat`] for given [`RpcTransport`].
     ///
-    /// Sends first `ping` immediately, so provided [`WebSocket`] must be
+    /// Sends first `ping` immediately, so provided [`RpcTransport`] must be
     /// active.
-    pub fn start(&self, socket: Rc<WebSocket>) -> Result<()> {
+    pub fn start(&self, transport: Rc<dyn RpcTransport>) -> Result<()> {
         let mut inner = self.0.borrow_mut();
         inner.num = 0;
         inner.pong_at = None;
-        inner.socket = Some(socket);
+        inner.transport = Some(transport);
         inner.send_now().map_err(tracerr::wrap!())?;
 
         let inner_rc = Rc::clone(&self.0);
@@ -117,7 +117,7 @@ impl Heartbeat {
     /// Stops [`Heartbeat`].
     pub fn stop(&self) {
         self.0.borrow_mut().ping_task.take();
-        self.0.borrow_mut().socket.take();
+        self.0.borrow_mut().transport.take();
     }
 
     /// Timestamp of last pong received.